parseCurrencyUnit & chopCurrencyUnitDecimals to be locale independant
- the problem is: we can't control if user input '.' or ',', we should accept both, and treat them both as a decimal separation
- that means we can't support anymore the thousands separator (aka grouping)
