+++ packages/currencies/src/parseCurrencyUnit.js
@@ -1,47 +1,13 @@
 //@flow
 import type { Unit } from "./types";
-import memoize from "lodash/memoize";
 
-const defaultFormatOptions = {
-  locale: "en-EN"
-};
-
-// returns decimal and thousands separator
-export const getSeparators = memoize((locale: string): {
-  decimal: ?string,
-  thousands: ?string
-} => {
-  const res = (10000.2).toLocaleString(locale);
-  let decimal, thousands;
-  for (let i = 0; i < res.length; i++) {
-    const c = res[i];
-    if (/[0-9]/.test(c)) continue;
-    if (!thousands) {
-      thousands = c;
-    } else {
-      decimal = c;
-    }
-  }
-  return { decimal, thousands };
-});
-
-export const parseCurrencyUnit = (
-  unit: Unit,
-  valueString: string,
-  options?: $Shape<typeof defaultFormatOptions>
-): number => {
-  const { locale } = {
-    ...defaultFormatOptions,
-    ...options
-  };
-  const sep = getSeparators(locale);
-  let str = "";
-  for (let i = 0; i < valueString.length; i++) {
-    let c = valueString[i];
-    if (c !== sep.thousands) {
-      str += c === sep.decimal ? "." : c;
-    }
-  }
+// parse a value that was formatted with formatCurrencyUnit
+// NB this function parse a subset of formats because it it locale independant.
+// make sure you have at least following options set on the formatter:
+// - useGrouping: true
+// - showCode: false
+export const parseCurrencyUnit = (unit: Unit, valueString: string): number => {
+  const str = valueString.replace(/,/g, ".");
   const value = parseFloat(str);
   if (isNaN(value)) return 0;
   return Math.round(value * 10 ** unit.magnitude);
